Don't inject content scripts with matchAboutBlank
For those frames, sender.url is about:blank, but if we don't get a real URL,
we cannot associate it to any credentials. If we are ever going to encounter
a login form in such a frame, we have to get the actual URL from the parent
frame or send it from the content script.

We previously raised the minimum_chrome_version to 39 for this feature.
Not exactly sure what the actual minimum required Chrome version is meanwhile,
but the newest feature I could identify was the menuItemId we check in the
contextMenus.onClicked handler, which was addedd in Chrome 35.